Just dump the view_location from the view to the buffer after each event.
Does what 2 previous commits did, but simpler and it does a bit more. It
allows a simple scenario. Imagine two panels opened (A and B), you move the
cursor in A and then open the buffer of A in B. It properly gets the view
location of A. NIH, fully replicates emacs behaviour.